requirements/development-requirements.txt
# load common requirements
-r common-requirements.txt

# standard imports
autopep8
coverage
black
pytest-django
pytest-factoryboy

# third party imports
bump2version

# RCPCH imports